+++ benchmarks/kernels/benchmark_paged_attention.py
@@ -5,7 +5,7 @@ from typing import Optional
 
 import torch
 
-from vllm.ops import ops
+from vllm import _custom_ops as ops
 from vllm.utils import STR_DTYPE_TO_TORCH_DTYPE, create_kv_caches_with_random
 
 NUM_BLOCKS = 1024
