+++ vllm/model_executor/layers/fused_moe/fused_moe.py
@@ -8,8 +8,8 @@ import torch
 import triton
 import triton.language as tl
 
+from vllm import _custom_ops as ops
 from vllm.logger import init_logger
-from vllm.ops import ops
 from vllm.utils import is_hip
 
 logger = init_logger(__name__)
