+++ vllm/model_executor/layers/quantization/marlin.py
@@ -3,11 +3,11 @@ from typing import Any, Dict, List, Optional
 import torch
 from torch.nn.parameter import Parameter
 
+from vllm import _custom_ops as ops
 from vllm.model_executor.layers.linear import (LinearMethodBase,
                                                set_weight_attrs)
 from vllm.model_executor.layers.quantization.base_config import (
     QuantizationConfig)
-from vllm.ops import ops
 
 
 class MarlinConfig(QuantizationConfig):
